Extract workspace thread mutation extension
