+++ Sources/QuillCodeApp/WorkspaceModelThreadMutation.swift
@@ -1,0 +1,61 @@
+import Foundation
+import QuillCodeCore
+
+@MainActor
+extension QuillCodeWorkspaceModel {
+    func appendNotice(_ summary: String) {
+        mutateSelectedThread { thread in
+            WorkspaceThreadNoticeAppender.appendNotice(summary, to: &thread)
+        }
+    }
+
+    func mutateSelectedThread(_ update: (inout ChatThread) -> Void) {
+        guard let selectedThreadID = root.selectedThreadID,
+              let index = mutateThread(selectedThreadID, update)
+        else {
+            return
+        }
+        root.selectedThreadID = root.threads[index].id
+        refreshTopBar(agentStatus: root.topBar.agentStatus)
+    }
+
+    func selectedSidebarThreadIDs() -> [UUID] {
+        let resolution = WorkspaceSidebarSelectionEngine.resolve(
+            state: sidebarSelection,
+            orderedSidebarThreadIDs: root.allSidebarItems.map(\.id),
+            validThreadIDs: validThreadIDs()
+        )
+        sidebarSelection = resolution.state
+        return resolution.selectedThreadIDs
+    }
+
+    func validThreadIDs() -> Set<UUID> {
+        Set(root.threads.map(\.id))
+    }
+
+    @discardableResult
+    func mutateThread(_ id: UUID, _ update: (inout ChatThread) -> Void) -> Int? {
+        guard let index = threadPersistence.mutate(id, threads: &root.threads, update: update) else {
+            return nil
+        }
+        refreshTopBar(agentStatus: root.topBar.agentStatus)
+        return index
+    }
+
+    func updateThreadFromAgentRun(_ thread: ChatThread) {
+        let result = WorkspaceThreadLifecycleEngine.applyAgentRunThreadUpdate(
+            thread,
+            threads: &root.threads,
+            projects: root.projects,
+            selectedThreadID: root.selectedThreadID,
+            selectedProjectID: root.selectedProjectID
+        )
+        root.selectedThreadID = result.selectedThreadID
+        root.selectedProjectID = result.selectedProjectID
+        if result.didSelectUpdatedThread {
+            syncTerminalSessionToSelectedProject()
+            touchProject(root.selectedProjectID)
+            saveProjects()
+        }
+    }
+}
