+++ Tests/QuillCodeParityTests/ParityWorkspaceSidebarGateTests.swift
@@ -4,6 +4,7 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
     func testWorkspaceModelDelegatesSidebarSelectionTransitions() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
         let threadExtensionText = try Self.appSourceText(named: "WorkspaceModelThreads.swift")
+        let threadMutationText = try Self.appSourceText(named: "WorkspaceModelThreadMutation.swift")
         let selectionText = try Self.appSourceText(named: "WorkspaceSidebarSelectionEngine.swift")
         let bulkPlannerText = try Self.appSourceText(named: "WorkspaceSidebarBulkActionPlanner.swift")
         let bulkExecutorText = try Self.appSourceText(named: "WorkspaceSidebarBulkActionExecutor.swift")
@@ -17,7 +18,7 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.start"), "WorkspaceModel thread/sidebar extension should delegate selection start.")
         XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.selectAll"), "WorkspaceModel thread/sidebar extension should delegate select-all.")
         XCTAssertTrue(threadExtensionText.contains("WorkspaceSidebarSelectionEngine.toggle"), "WorkspaceModel thread/sidebar extension should delegate selection toggles.")
-        XCTAssertTrue(modelText.contains("WorkspaceSidebarSelectionEngine.resolve"), "WorkspaceModel should delegate stale-ID pruning and ordering.")
+        XCTAssertTrue(threadMutationText.contains("WorkspaceSidebarSelectionEngine.resolve"), "WorkspaceModel thread mutation extension should delegate stale-ID pruning and ordering.")
         XCTAssertTrue(bulkPlannerText.contains("struct WorkspaceSidebarBulkActionPlanner"), "Sidebar bulk action planning should live in a focused planner.")
         XCTAssertTrue(bulkPlannerText.contains("static func plan"), "Sidebar bulk action plans should be directly testable.")
         XCTAssertTrue(bulkPlannerText.contains("enum FollowUpSelection"), "Bulk action selection follow-up policy should be explicit.")
@@ -31,6 +32,7 @@ final class ParityWorkspaceSidebarGateTests: QuillCodeParityTestCase {
         XCTAssertFalse(modelText.contains("selectedThreadIDs.insert"), "WorkspaceModel should not mutate sidebar selection sets directly.")
         XCTAssertFalse(modelText.contains("selectedThreadIDs.remove"), "WorkspaceModel should not mutate sidebar selection sets directly.")
         XCTAssertFalse(modelText.contains("selectedThreadIDs.intersection"), "WorkspaceModel should not prune sidebar selection sets directly.")
+        XCTAssertFalse(modelText.contains("WorkspaceSidebarSelectionEngine.resolve"), "WorkspaceModel.swift should not own stale-ID pruning and ordering.")
         XCTAssertFalse(modelText.contains("let ids = selectedSidebarThreadIDs()"), "WorkspaceModel should not inline bulk selected-ID planning.")
         XCTAssertFalse(modelText.contains("case .pin(let ids):"), "WorkspaceModel should not execute sidebar bulk pin mutations inline.")
         XCTAssertFalse(modelText.contains("WorkspaceThreadLifecycleEngine.archiveThreads"), "WorkspaceModel should not execute sidebar bulk archive mutations inline.")
